chore: 🔧 unify server actions under /app/actions directory
