+++ src/db/schema.ts
@@ -135,7 +135,9 @@ export const meetings = pgTable("meetings", {
     group_id: uuid("group_id").references(() => groups.id, {
         onDelete: "cascade",
     }),
-    host_id: text("host_id").references(() => members.id).notNull(),
+    host_id: text("host_id")
+        .references(() => members.id)
+        .notNull(),
     // dates: interval("dates"), -- STORES RELATIVE TIME INTERVAL
     // dates: jsonb("dates") -- CANNOT VALIDATE KEY-VALUE PAIR FORMAT IN DDL
     dates: jsonb("dates").notNull().default([]),
@@ -150,24 +152,6 @@ export const meetingsRelations = relations(meetings, ({ one, many }) => ({
     }),
     availabilities: many(availabilities),
 }));
-
-// TODO: remove this table
-/**
- * @deprecated in favor of storing dates in the meetings table as JSON column
- */
-// export const meetingDates = pgTable(
-//     "meeting_dates",
-//     {
-//         id: uuid("id").unique().defaultRandom(),
-//         meeting_id: uuid("meeting_id").references(() => meetings.id, {
-//             onDelete: "cascade",
-//         }),
-//         date: timestamp("date").notNull(),
-//     },
-//     (table) => ({
-//         pk: primaryKey({ columns: [table.id, table.date] }),
-//     })
-// );
 
 export const groups = pgTable("groups", {
     id: uuid("id").defaultRandom().primaryKey(),
